move `/colors` to the root
