+++ src/routes/+page.svelte
@@ -1,7 +1,28 @@
 <script lang="ts">
-	import { goto } from '$app/navigation';
+	import { colord, type Colord } from 'colord';
 
-	$effect.pre(() => {
-		goto('/colors');
-	});
+	import { generateRandomColor } from '$lib/utils';
+
+	import ColorPaletteCard from './ColorPaletteCard.svelte';
+
+	const colors: Colord[] = $state([]);
 </script>
+
+<h1>Explore and Create Color Combos</h1>
+<p>A creative tool to experiment with and perfect your color combinations.</p>
+<button onclick={() => colors.push(colord(generateRandomColor()))}>New color</button>
+
+<div class="color-grid">
+	{#each colors as color}
+		<ColorPaletteCard {color} />
+	{/each}
+</div>
+
+<style>
+	.color-grid {
+		margin-top: 20px;
+		display: grid;
+		grid-template-columns: repeat(auto-fill, 300px);
+		gap: 20px;
+	}
+</style>
