+++ src/routes/ColorPaletteCard.svelte
